move watch.ListWatchUntil to its own package to avoid future import cycle
