+++ pkg/watch/BUILD
@@ -22,8 +22,6 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
-        "//pkg/api:go_default_library",
-        "//pkg/api/meta:go_default_library",
         "//pkg/api/unversioned:go_default_library",
         "//pkg/runtime:go_default_library",
         "//pkg/util/net:go_default_library",
